pyproject.toml
[project]
name = "nyl"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Niklas Rosenstein", email = "rosensteinniklas@gmail.com"},
]
dependencies = [
    "databind>=4.5.2",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.4",
    "loguru>=0.7.2",
    "kubernetes>=30.1.0",
    "typer>=0.12.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
nyl = "nyl.main:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.4.8",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0.12.20240311",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.28",
    "mkdocstrings[python]>=0.25.1",
]

[tool.ruff]
line-length = 120
